Move logic from dom to xml module
