+++ src/ol/xml.js
@@ -2,7 +2,6 @@
  * @module ol/xml
  */
 import {extend} from './array.js';
-import {getDocument} from './dom.js';
 
 
 /**
@@ -508,4 +507,29 @@ export function getXMLSerializer() {
     xmlSerializer_ = new XMLSerializer();
   }
   return xmlSerializer_;
+}
+
+
+let document_ = undefined;
+
+/**
+ * Register a Document to use when creating nodes for XML serializations. Can be used
+ * to inject a Document where there is no globally available implementation.
+ *
+ * @param {Document} document A Document.
+ * @api
+ */
+export function registerDocument(document) {
+  document_ = document;
+}
+
+/**
+ * Get a document that should be used when creating nodes for XML serializations.
+ * @return {Document} The document.
+ */
+export function getDocument() {
+  if (document_ === undefined && typeof document !== 'undefined') {
+    document_ = document.implementation.createDocument('', '', null);
+  }
+  return document_;
 }
